Convert Tips resource to tsoa
